Renamed IssuePanel to ListPanel
