+++ src/test/java/guitests/KeyboardShortcutsTest.java
@@ -4,7 +4,7 @@ import javafx.scene.input.KeyCode;
 import org.junit.Test;
 import ui.UI;
 import ui.components.KeyboardShortcuts;
-import ui.issuepanel.IssuePanel;
+import ui.listpanel.ListPanel;
 import util.events.ColumnClickedEventHandler;
 import util.events.IssueSelectedEventHandler;
 import util.events.testevents.UIComponentFocusEvent;
@@ -99,7 +99,7 @@ public class KeyboardShortcutsTest extends UITest {
         click("#dummy/dummy_col1_1");
 
         // mark as read/unread
-        IssuePanel issuePanel = find("#dummy/dummy_col1");
+        ListPanel issuePanel = find("#dummy/dummy_col1");
         push(getKeyCode("MARK_AS_READ"));
         assertEquals(true, issuePanel.getSelectedIssue().isCurrentlyRead());
         push(getKeyCode("MARK_AS_UNREAD"));
