+++ src/test/java/guitests/RangeTest.java
@@ -3,7 +3,7 @@ package guitests;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
 import org.junit.Test;
-import ui.issuepanel.IssuePanel;
+import ui.listpanel.ListPanel;
 
 import static org.junit.Assert.assertEquals;
 
@@ -17,7 +17,7 @@ public class RangeTest extends UITest {
         click("#dummy/dummy_col0_filterTextField");
         press(KeyCode.ENTER).release(KeyCode.ENTER);
         sleep(EVENT_DELAY);
-        assertEquals(5, ((IssuePanel) find("#dummy/dummy_col0")).getIssueCount());
+        assertEquals(5, ((ListPanel) find("#dummy/dummy_col0")).getIssueCount());
     }
 
     @Test
@@ -26,6 +26,6 @@ public class RangeTest extends UITest {
         click("#dummy/dummy_col0_filterTextField");
         press(KeyCode.ENTER).release(KeyCode.ENTER);
         sleep(EVENT_DELAY);
-        assertEquals(6, ((IssuePanel) find("#dummy/dummy_col0")).getIssueCount());
+        assertEquals(6, ((ListPanel) find("#dummy/dummy_col0")).getIssueCount());
     }
 }
